+++ packages/@uppy/utils/package.json
@@ -27,7 +27,6 @@
     "./lib/dataURItoBlob": "./lib/dataURItoBlob.js",
     "./lib/dataURItoFile": "./lib/dataURItoFile.js",
     "./lib/emaFilter": "./lib/emaFilter.js",
-    "./lib/emitSocketProgress": "./lib/emitSocketProgress.js",
     "./lib/findAllDOMElements": "./lib/findAllDOMElements.js",
     "./lib/findDOMElement": "./lib/findDOMElement.js",
     "./lib/generateFileID": "./lib/generateFileID.js",
